server/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-login
python-dotenv
google-generativeai
google-genai
